requirements.txt
streamlit==1.50.0
openai-agents==0.4.2
langchain-community==0.4.1
pypdf==6.1.3
google-search-results

python-dotenv~=1.2.1
pydantic~=2.12.3
fitz~=0.0.1.dev2
pillow~=11.3.0
paddleocr~=3.3.0
